--- a/crates/llama-cpp-bindings/include/engine.h
+++ b/crates/llama-cpp-bindings/include/engine.h
@@ -11,6 +11,7 @@ class TextInferenceEngine {
 
   virtual uint32_t start(const rust::Str prompt) const = 0;
   virtual uint32_t step(uint32_t next_token_id) const = 0;
+  virtual void end() const = 0;
 
   virtual uint32_t eos_token() const = 0;
 };