@wsxiaoys 2023-09-03
@ggerganov conducted an experimental integration with ggml's metal backend and produced the results shown above. Do you think the [engine.cc](https://github.com/TabbyML/tabby/pull/393/files#diff-117f03b0d5d452660f112a44de48d900dabbb06a7725b1759fa0c60df90745df) overall looks good? Please let me know if there are any flags worth trying to improve performance.