@wsxiaoys 2023-09-03
@ggerganov did an experimental integration with ggml's metal backend and generate numbers above. Does the  [engine.cc](https://github.com/TabbyML/tabby/pull/393/files#diff-117f03b0d5d452660f112a44de48d900dabbb06a7725b1759fa0c60df90745df) overall looks good to you? Let me know if there's some worth to try flags for better performance.