@ggerganov 2023-09-05
Do you have `llama.cpp` version from yesterday?
There were some significant performance improvements for F16 on Metal that would give you >20 t/s for M2 Max for 7B.

`engine.cc` looks OK

Let me know if you have more questions