@ggerganov 2023-09-05
Yes, seems to be included.

On M2 Ultra, I get ~40 t/s with Code Llama 7B F16, so I was expecting at least ~20 t/s on M2 Max:

```bash
./bin/main -m ../models/codellama-7b/ggml-model-f16.gguf -p "# Dijkstra's shortest path algorithm in Python (4 spaces indentation) + complexity analysis:\n\n" -e -t 4 --temp -1 -n 128

system_info: n_threads = 4 / 24 | AVX = 0 | AVX2 = 0 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 0 | NEON = 1 | ARM_FMA = 1 | F16C = 0 | FP16_VA = 1 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 0 | SSSE3 = 0 | VSX = 0 | 
sampling: repeat_last_n = 64, repeat_penalty = 1.100000, presence_penalty = 0.000000, frequency_penalty = 0.000000, top_k = 40, tfs_z = 1.000000, top_p = 0.950000, typical_p = 1.000000, temp = -1.000000, mirostat = 0, mirostat_lr = 0.100000, mirostat_ent = 5.000000
generate: n_ctx = 512, n_batch = 512, n_predict = 128, n_keep = 0


 # Dijkstra's shortest path algorithm in Python (4 spaces indentation) + complexity analysis:

### Complexity Analysis:

- Time Complexity: O(V^2), where V is the number of vertices.
- Space Complexity: O(V).

### Pseudocode:

```python
Dijkstra(Graph, source):
    dist[source] = 0                        // Distance from source to source
    for each vertex v in Graph:              // Initialization
        if v != source:
            dist[v] = INFINITY               // Unknown distance function from source to v
            prev[v] = UNDEFINED
llama_print_timings:        load time =   540.25 ms
llama_print_timings:      sample time =    81.64 ms /   128 runs   (    0.64 ms per token,  1567.76 tokens per second)
llama_print_timings: prompt eval time =    45.21 ms /    25 tokens (    1.81 ms per token,   552.93 tokens per second)
llama_print_timings:        eval time =  3180.65 ms /   127 runs   (   25.04 ms per token,    39.93 tokens per second)
llama_print_timings:       total time =  3328.93 ms
ggml_metal_free: deallocating
Log end
```