@wsxiaoys 2023-09-05
```
# assume there should be `-ngl 1`

./bin/main -m ../models/codellama-model-f16.gguf -p "# Dijkstra's shortest path algorithm in Python (4 spaces indentation) + complexity analysis:\n\n" -e -t 4 --temp -1 -n 128 -ngl 1

system_info: n_threads = 4 / 12 | AVX = 0 | AVX2 = 0 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 0 | NEON = 1 | ARM_FMA = 1 | F16C = 0 | FP16_VA = 1 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 0 | SSSE3 = 0 | VSX = 0 |
sampling: repeat_last_n = 64, repeat_penalty = 1.100000, presence_penalty = 0.000000, frequency_penalty = 0.000000, top_k = 40, tfs_z = 1.000000, top_p = 0.950000, typical_p = 1.000000, temp = -1.000000, mirostat = 0, mirostat_lr = 0.100000, mirostat_ent = 5.000000
generate: n_ctx = 512, n_batch = 512, n_predict = 128, n_keep = 0


 # Dijkstra's shortest path algorithm in Python (4 spaces indentation) + complexity analysis:

### Complexity Analysis:

- Time Complexity: O(V^2), where V is the number of vertices.
- Space Complexity: O(V).

### Pseudocode:

```python
Dijkstra(Graph, source):
    dist[source] = 0                        // Distance from source to source
    for each vertex v in Graph:              // Initialization
        if v != source:
            dist[v] = INFINITY               // Unknown distance function from source to v
            prev[v] = UNDEFINED
llama_print_timings:        load time =   997.82 ms
llama_print_timings:      sample time =    82.49 ms /   128 runs   (    0.64 ms per token,  1551.74 tokens per second)
llama_print_timings: prompt eval time =    63.77 ms /    25 tokens (    2.55 ms per token,   392.03 tokens per second)
llama_print_timings:        eval time =  5228.16 ms /   127 runs   (   41.17 ms per token,    24.29 tokens per second)
llama_print_timings:       total time =  5396.12 ms
ggml_metal_free: deallocating
Log end
```

The performance is indeed ~41ms per token (24t / second). 

Let me compare the engine.cc and example/main.cc in depth...