@wsxiaoys 2023-09-05
I just realized that I updated the PR's llama.cpp version in commit https://github.com/TabbyML/tabby/pull/393/commits/8d01604cfd650fc8a69b8e6aa048e1e5137aa422 without redoing the measurement

Here are the statistics for the latest version:

## float16

```bash
llama_print_timings:        load time =  4544.11 ms
llama_print_timings:      sample time =     0.00 ms /     1 runs   (    0.00 ms per token,      inf tokens per second)
llama_print_timings: prompt eval time =   464.48 ms /    29 tokens (   16.02 ms per token,    62.44 tokens per second)
llama_print_timings:        eval time =  1176.45 ms /    29 runs   (   40.57 ms per token,    24.65 tokens per second)
llama_print_timings:       total time =  1642.56 ms
```

## q8_0
```bash
llama_print_timings:        load time =   293.14 ms
llama_print_timings:      sample time =     0.00 ms /     1 runs   (    0.00 ms per token,      inf tokens per second)
llama_print_timings: prompt eval time =   292.94 ms /    29 tokens (   10.10 ms per token,    99.00 tokens per second)
llama_print_timings:        eval time =   709.46 ms /    29 runs   (   24.46 ms per token,    40.88 tokens per second)
llama_print_timings:       total time =  1003.70 ms
```

I am amazed by the 2x performance improvement. Thanks for pointing it out!